@@ -1,5 +1,9 @@
 import React from "react";
-import ReactDOM from "react-dom";
-import App from "./components/App";
+import ReactDOM from 'react-dom/client';
+import App from "./App";
+import searchImages from './api';
 
-ReactDOM.render(<App />, document.querySelector("#root"));
+const el = document.getElementById('root');
+const root = ReactDOM.createRoot(el);
+
+root.render(<App />);
